feat: Update mock data and schema for Claim and QuoteBlock
This commit introduces changes to integration tests' mock data
by adding a 'Claim' and a 'QuoteBlock'. Corresponding test cases
are added in http-schema.test.ts. The http-schema.ts is updated
to modify the claimType in ClaimContentSchema to be optional
and add a new content field. Also, the 'QuoteBlock' option
is added to the OtherContent object.