feat: allow volumeNumber to be safe number or string (#850)
* feat: update volumeNumber field validation and type

Update the type of the volumeNumber field from number to string in the
Publication type. Also, modify the validation for the volumeNumber field in
the PublicationSchema to a custom validation that checks if the value is a
number. Add volumeNumber field in the test data for the http-schema tests.

* feat: allow volumeNumber to be number or string

Updated the 'volumeNumber' field in the 'Publication' model to accept both
numbers and strings. Adjusted Joi validation in http-schema to handle this
change and convert unsafe numbers to strings. Updated http-schema tests
accordingly to reflect this change.