feat: Update mock data and schema for Claim and QuoteBlock (#851)
* feat: Update mock data and schema for Claim and QuoteBlock

This commit introduces changes to integration tests' mock data
by adding a 'Claim' and a 'QuoteBlock'. Corresponding test cases
are added in http-schema.test.ts. The http-schema.ts is updated
to modify the claimType in ClaimContentSchema to be optional
and add a new content field. Also, the 'QuoteBlock' option
is added to the OtherContent object.

* feat: make claimType optional in ClaimContent model

This commit updates the ClaimContent model in 'src/model/content.ts' by making
the 'claimType' field optional. Corresponding tests have also been updated
in 'src/http-schema/http-schema.test.ts' to reflect this change.

* feat(http-schema): make content field required in http-schema

- Added test for 'content' field in http-schema.test.ts
- Changed 'content' field from optional to required in http-schema.ts