feat: add pagination and sorting to getEnhancedArticlesNoContent (#920)
* feat: add pagination and sorting to getEnhancedArticlesNoContent

The getEnhancedArticlesNoContent method in the Preprints Controller, Model, and MongoDB Repository now supports pagination and sorting. It accepts 'order' (asc/desc), 'page', and 'per-page' parameters and provides a default value when these parameters are not provided. The MongoDB query has been updated with aggregate pipeline operators for matching, sorting, and grouping the results.

* feat: Add total count to enhanced preprints response

This commit introduces a new feature to the enhanced preprints server.
It adds a total count of enhanced articles without content to the
response header. This is done by adding a new method in the repository
to get the total count, and setting that count in the response header
in the preprints controller. Also, an export type is added to the model
to handle the total count.

* feat: update API tests for preprints with no content

Adjusted the expectations for the API endpoint '/api/preprints-no-content' in the integration tests.
The number of returned items is now verified against the 'X-Total-Count' header and the body length.
Removed assertions related to the second item returned by this endpoint.

* feat: change method in api test to get X-Total-Count from response header

* feat: update header validation in integration tests

In the api.test.ts file, the validation of the 'X-Total-Count' header in the
/api/preprints-no-content response has been updated. Now, the test first checks if the
header is present before verifying its value.

* feat: Remove x-total-count header check in api test

The check for the presence of 'x-total-count' header in the api test file
has been removed. The value of this header is still being validated.

* feat: add additional example versions in api integration tests

This commit adds three new example versions in the integration tests for the
API. The tests now check for correct responses when posting these new versions
and getting the updated list of preprints. The tests also cover pagination
and ordering functionality of the 'preprints-no-content' API endpoint.