feat: Add first published date to EnhancedArticleNoContent (#923)
* feat: Add first published date to EnhancedArticleNoContent

This commit modifies the integration tests and MongoDB repository to include the first published date in the `EnhancedArticleNoContent` model. The date is determined by finding the minimum 'published' value for each document. Tests have been updated to verify this new functionality.

* feat: remove exclusive test in api test file

The `only` method was removed from a test in the `api.test.ts` file to allow all tests to run.