[#1005] feat: add end-date filtering for fetching versions without content
This commit includes the addition of 'end-date' query parameter for
the endpoint '/api/preprints-no-content'. The endpoint now supports
filtering of versions of articles that do not have content by the
specified end-date.

The integration test for this feature is updated accordingly and is
passing. Changes are also made in 'preprints-controller.ts',
'mongodb-repository.ts', and 'model.ts' to accommodate this feature.