Add meta properties to Article, Organisation and Author models [elifesciences/enhanced-preprints-issues#1159]
This commit includes updates to the 'model.ts' file to include the 'meta' property to the 'Author', 'Organisation', and 'Article' models. This property supports metadata for authors and organisations, as well as article metadata.

The 'http-schema.ts' file has been updated to include validation for the new 'meta' property.

The 'http-schema.test.ts' has been updated to include tests for the new metadata validation rules.